chore(requests): Make test utils available in getsentry
